crate::fraction module refactoring; support for transparent operand cast with basic math ops
